blog_if/blog/models.py
# Importações necessárias
from ckeditor.fields import RichTextField
from ckeditor_uploader.fields import RichTextUploadingField
from django.contrib.auth.models import User
from django.db import models

# Modelo Post que representa um post do blog dentro do banco de dados
class Post(models.Model):
    title = models.CharField(max_length=255)
    summary = RichTextField()
    content = RichTextUploadingField()
    author = models.ForeignKey(User, on_delete=models.PROTECT)
    created_at = models.DateTimeField(auto_now_add=True)

# Método que retorna o título do post da forma que ele deve ser exibido
    def __str__(self):
        return self.title