blog_if/requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.1.0
filelock==3.13.1
Flask==2.3.2
fonttools==4.42.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==4.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.25.0
Werkzeug==2.3.6
